# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GigSecure",
    "description": "Digital insurance platform for Africa's gig economy",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh Token",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" } } }, "required": true },
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": { "description": "Validation Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": { "properties": { "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array" } }, "type": "object", "title": "HTTPValidationError" },
...
      "RefreshTokenRequest": { "properties": { "refresh_token": { "type": "string" } }, "type": "object", "required": ["refresh_token"], "title": "RefreshTokenRequest" },
...
      "ValidationError": { "properties": { "loc": { "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }, "type": "array" }, "msg": { "type": "string" }, "type": { "type": "string" } }, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError" },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}